+++ timeMux/bddMux.cpp
@@ -18,7 +18,7 @@ void buildTrans(DdManager *dd, DdNode **pNodeVec, DdNode **B, cuint nPi, cuint n
 
     st__generator *cGen, *nGen;
     DdNode *cKNode, *cVNode, *nKNode, *nVNode;
-    size_t cCnt = 0, nCnt = 0;
+    uint cCnt = 0, nCnt = 0;
     
     DdNode **oFuncs  = (i==nTimeFrame-1) ? new DdNode*[nCo] : NULL;
     DdNode *F, *G, *path, *bCube, *tmp1, *tmp2;
@@ -30,7 +30,7 @@ void buildTrans(DdManager *dd, DdNode **pNodeVec, DdNode **B, cuint nPi, cuint n
         // get output function of current state
         // oFunc = cofactor(f_k, cube)
         bCube = Cudd_CubeArrayToBdd(dd, cube);  Cudd_Ref(bCube);  // bdd of the cube
-        if(oFuncs) for(size_t k=0; k<nCo; ++k) {
+        if(oFuncs) for(uint k=0; k<nCo; ++k) {
             oFuncs[k] = Cudd_Cofactor(dd, pNodeVec[k], bCube);
             Cudd_Ref(oFuncs[k]);
         }
@@ -40,7 +40,7 @@ void buildTrans(DdManager *dd, DdNode **pNodeVec, DdNode **B, cuint nPi, cuint n
         // encode output on/off-sets
         if(oFuncs) {
             F = b0;  Cudd_Ref(F);
-            for(size_t k=0; k<2; ++k) {
+            for(uint k=0; k<2; ++k) {
                 bddNotVec(oFuncs, nCo);   // 2*negation overall
                 tmp1 = bddDot(dd, oFuncs, B+k*nCo, nCo);
                 tmp2 = Cudd_bddOr(dd, F, tmp1);  Cudd_Ref(tmp2);
@@ -281,9 +281,9 @@ DdManager* buildGlobalBdd(Abc_Ntk_t *pNtk, int nBddSizeMax, int fReorder, int in
 int buildAigHyper(Abc_Ntk_t *&pNtk, cuint optRnd = 3)
 {
     int i; Abc_Obj_t *pObj;
-    size_t nCo = Abc_NtkCoNum(pNtk);
-    size_t nCi = Abc_NtkCiNum(pNtk);
-    size_t nv = nCo ? (size_t)ceil(log2(double(nCo))) : 0;
+    uint nCo = Abc_NtkCoNum(pNtk);
+    uint nCi = Abc_NtkCiNum(pNtk);
+    uint nv = nCo ? (uint)ceil(log2(double(nCo))) : 0;
     
     Abc_Obj_t **B = aigUtils::aigComputeSign(pNtk, nCo, false);
     assert(nCi + nv == Abc_NtkCiNum(pNtk));
@@ -344,7 +344,7 @@ int bddMux(Abc_Ntk_t *pNtk, cuint nTimeFrame, vector<string>& stg, int *perm, co
     assert((j == nCo) && H);
 
     // compute signature
-    size_t nB = nCo ? (size_t)ceil(log2(double(nCo*2))) : 0;
+    uint nB = nCo ? (uint)ceil(log2(double(nCo*2))) : 0;
     DdNode **B = bddComputeSign(dd, nCo*2, initVarSize);
     assert(initVarSize+nB == Cudd_ReadSize(dd));
     if(logger) logger->log("prepare sign");
@@ -381,7 +381,7 @@ int bddMux(Abc_Ntk_t *pNtk, cuint nTimeFrame, vector<string>& stg, int *perm, co
         pNodeVec[i] = (DdNode *)Abc_ObjGlobalBdd(pObj);  // ref = 1
 
     // compute signature
-    size_t nB = nCo ? (size_t)ceil(log2(double(nCo*2))) : 0;
+    uint nB = nCo ? (uint)ceil(log2(double(nCo*2))) : 0;
     DdNode **B = bddComputeSign(dd, nCo*2);
     assert(initVarSize+nB == Cudd_ReadSize(dd));
     if(logger) logger->log("prepare sign");
@@ -425,7 +425,7 @@ int bddMux(Abc_Ntk_t *pNtk, cuint nTimeFrame, vector<string>& stg, int *perm, co
 
 
     // collecting states at each timeframe
-    size_t stsSum = 1;
+    uint stsSum = 1;
     st__table *csts = bddCreateDummyState(dd), *nsts;
     
     for(i=0; i<nTimeFrame; ++i) {
