+++ timeMux/checkEqv.cpp
@@ -18,12 +18,12 @@ static inline void dumpSTG(Abc_Ntk_t *pNtk, cuint nTimeFrame, cuint nSts, const 
     cuint nPi = Abc_NtkCiNum(pNtk) / nTimeFrame;
     cuint nPo = Abc_NtkCoNum(pNtk);
     
-    sprintf(buf1, "%lu-stg.kiss", pid);
+    sprintf(buf1, "%u-stg.kiss", pid);
     ofstream fp = ofstream(buf1);
     fileWrite::writeKiss(nPi, nPo, nSts, stg, fp);
     fp.close();
 
-    sprintf(buf2, "./MeMin %s %lu-mstg.kiss > /dev/null", buf1, pid);
+    sprintf(buf2, "./MeMin %s %u-mstg.kiss > /dev/null", buf1, pid);
     system(buf2);
     remove(buf1);
 }
@@ -32,8 +32,8 @@ static inline void dumpSTG(Abc_Ntk_t *pNtk, cuint nTimeFrame, cuint nSts, const 
 static inline void dumpBLIF(cuint pid)
 {
     char buf1[100], buf2[1000];
-    sprintf(buf1, "%lu-mstg.kiss", pid);
-    sprintf(buf2, "./kiss2blif %s %lu-mstg.blif", buf1, pid);
+    sprintf(buf1, "%u-mstg.kiss", pid);
+    sprintf(buf2, "./kiss2blif %s %u-mstg.blif", buf1, pid);
     system(buf2);
     remove(buf1);
 }
@@ -43,7 +43,7 @@ static Abc_Ntk_t* prepNtkToCheck(cuint nTimeFrame, int *perm, cuint pid)
 {
     // read the BLIF file
     char buf[100];
-    sprintf(buf, "%lu-mstg.blif", pid);
+    sprintf(buf, "%u-mstg.blif", pid);
     Abc_Ntk_t *pNtk = Io_Read(buf, Io_ReadFileType(buf), 1, 0);
     remove(buf);
 
@@ -56,7 +56,7 @@ static Abc_Ntk_t* prepNtkToCheck(cuint nTimeFrame, int *perm, cuint pid)
     Abc_NtkDelete(pNtkStr);
 
     // retrieve the outputs from the last time-frame
-    size_t nPo = Abc_NtkCoNum(pNtk) / nTimeFrame;
+    uint nPo = Abc_NtkCoNum(pNtk) / nTimeFrame;
     pNtkStr = aigUtils::aigCone(pNtk, Abc_NtkCoNum(pNtk)-nPo, Abc_NtkCoNum(pNtk));
     Abc_NtkDelete(pNtk);
 
@@ -70,14 +70,14 @@ static Abc_Ntk_t* prepNtkToCheck(cuint nTimeFrame, int *perm, cuint pid)
 
 void checkEqv(Abc_Ntk_t *pNtk, int *perm, cuint nTimeFrame, const vector<string>& stg, cuint nSts)
 {
-    size_t pid = (size_t)getpid();  // name the tmp dump files with pid to avoid name collision
+    uint pid = (uint)getpid();  // name the tmp dump files with pid to avoid name collision
 
     dumpSTG(pNtk, nTimeFrame, nSts, stg, pid);
     dumpBLIF(pid);
     Abc_Ntk_t *pNtkCheck = prepNtkToCheck(nTimeFrame, perm, pid);
     
     //cout << "prep done, start checking..." << endl;
-    //if(perm) for(size_t i=0; i<Abc_NtkCiNum(pNtk); ++i)
+    //if(perm) for(uint i=0; i<Abc_NtkCiNum(pNtk); ++i)
     //    cout << i << " -> " << perm[i] << endl;;
 
     // perform CEC and print message
